LayoutTests/fast/viewport/viewport-6-expected.txt

ALERT: viewport size 200x220 scale 1.6 with limits [1.6, 5] and userScalable true
